ai/mlops_project/infra/airflow/requirements.txt
pendulum
numpy
pandas
scikit-learn
joblib
psycopg2-binary
mlflow
